--- geosnap/data/__init__.py
@@ -1,4 +1,0 @@
-from .data import Community, store_ltdb, store_ncdb, store_census, data_store
-
-
-from .util import convert_gdf, get_lehd, adjust_inflation
