+++ geosnap/io/variables.csv
